Refactor and sendNotifcation fix
