+++ Source/ScaleStructureController/ScaleDesignWindow.h
